garim_attack_server/requirements.txt
# --- Real-time Communication ---
python-socketio==5.11.0
eventlet==0.33.3
Flask==3.0.0
Flask-Cors==4.0.0

# --- AI & Image Processing ---
insightface==0.7.3
onnxruntime==1.16.3
opencv-python==4.8.1.78
numpy==1.26.2

# --- Utilities ---
requests==2.31.0
python-dotenv==1.0.0